backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pypdf
python-docx
httpx
python-dotenv
ollama
